# Improved Razor formatting performance and added a perf benchmark