… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/LanguageServer/RazorCSharpFormattingBenchmark.cs
… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/LanguageServer/RazorLanguageServerBenchmarkBase.cs
… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/Microsoft.AspNetCore.Razor.Performance.csproj
… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/ProjectSystem/BackgroundCodeGenerationBenchmark.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Formatting/HtmlFormatter.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Properties/AssemblyInfo.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/RazorLanguageServer.cs
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting/CodeDirectiveFormattingTest.cs
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting/CodeDirectiveOnTypeFormattingTest.cs
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting/HtmlFormattingTest.cs
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting/TestFiles/HtmlFormattingTest/FormatsMixedRazorBlock.input.html
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting/TestFiles/HtmlFormattingTest/FormatsMixedRazorBlock.output.html
… src/Razor/test/Microsoft.AspNetCore.Razor.OmniSharpPlugin.Test/Microsoft.AspNetCore.Razor.OmniSharpPlugin.Test.csproj
… src/Razor/test/testapps/ComponentApp/Components/Pages/FormattingTest.razor