#### Summary of the changes
- Redesigned formatting to only invoke C# formatter once
- Added a RazorCSharpFormattingBenchmark
- Updated Roslyn version to consume a bug fix
- Fixed up some analyzer warnings

Fixes: https://github.com/dotnet/aspnetcore/issues/27144

#### Benchmark results:
The numbers below were obtained for formatting a [200 line Razor file](https://github.com/dotnet/aspnetcore-tooling/blob/efa7ab86fb6e56a2325396b328d3df6d876142a3/src/Razor/test/testapps/ComponentApp/Components/Pages/FormattingTest.razor#L1) with a fair number of Html->C#->Html transitions.

Before:

|                    Method |    Mean |    Error |   StdDev |   Op/s |      Gen 0 |      Gen 1 | Gen 2 | Allocated |
|-------------------------- |--------:|---------:|---------:|-------:|-----------:|-----------:|------:|----------:|
| 'Razor CSharp Formatting' | 6.676 s | 0.0743 s | 0.0695 s | 0.1498 | 37000.0000 | 10000.0000 |     - |   2.78 GB |


After:

|                    Method |     Mean |   Error |  StdDev |  Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|--------:|--------:|------:|------:|------:|------:|----------:|
| 'Razor CSharp Formatting' | 109.3 ms | 1.06 ms | 0.99 ms | 9.152 |     - |     - |     - |  42.67 MB |