@ajaybhargavb 2020-12-08
> Totally out of curiosity how does formatting performance scale? Aka, if you take a 400 line file does it take 200ms to compute?

Sort of. 

~400 lines,
|                    Method |     Mean |   Error |   StdDev |  Op/s |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|--------:|---------:|------:|----------:|------:|------:|----------:|
| 'Razor CSharp Formatting' | 262.0 ms | 6.06 ms | 17.49 ms | 3.817 | 1000.0000 |     - |     - |  96.24 MB |


~800 lines,
|                    Method |     Mean |    Error |   StdDev |  Op/s |     Gen 0 |     Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|---------:|---------:|------:|----------:|----------:|------:|----------:|
| 'Razor CSharp Formatting' | 688.7 ms | 13.70 ms | 39.10 ms | 1.452 | 3000.0000 | 1000.0000 |     - | 243.92 MB |

There is definitely room for improvement. I'm sure we'll find a whole bunch of quick wins when we get to actually profiling this code path.

Another point to note is that the file I used for these benchmarks have a lot of C#/HTML transitions (aka lot of source mappings). On average, the number of source mappings should still be reasonable even if the lines of code is higher.