@NTaylorMullen 2020-12-08
> There is definitely room for improvement. I'm sure we'll find a whole bunch of quick wins when we get to actually profiling this code path.
> 
> Another point to note is that the file I used for these benchmarks has a lot of C#/HTML transitions (aka lot of source mappings). On average, the number of source mappings should still be reasonable even if the lines of code is higher.

Ya totally fair. It'll probably be something we'll need to invest in at some point but we'll let users complain 😄